fix(fe): refactor lib folder
