+++ apps/frontend/components/DataTable.tsx
@@ -8,7 +8,7 @@ import {
   TableHeader,
   TableRow
 } from '@/components/ui/table'
-import { cn } from '@/lib/utils'
+import { cn } from '@/libs/utils'
 import type { ColumnDef } from '@tanstack/react-table'
 import {
   flexRender,
